Quantizing Procedure
Finished 'forward' method - the intermediary between the encoder and decoder, where the quantization of continuous data from the encoder happens by assigning it to its closest vector in the codebook. Loss was also implemented to enforce  encoder/codebook harmony. This function is also responsible for sending the quantized vector to the decoder for further processing.